+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Bug fixes
 
-- `eta_squared(<afex> / <Anova.mlm>)` with partially overlaping factor names now return correct effect sizes ( #697 )
+- `eta_squared(<afex> / <Anova.mlm>)` with partially overlapping factor names now return correct effect sizes ( #697 )
 
 # effectsize 1.0.2